move `ReadResponseBody` function to util
